==> naman_bhandari_capstone/notebooks/01-Preprocessing.ipynb <==
# Pre-processing data

In this notebook I compressed my CSV files by casting object columns as categories. This reduces Pandas' memory usage by hashing data in each column. I use the following libraries:

1. `pandas` - A library for manipulating data in a tabular format
2. `pickle` - used to store data for later use

I created this notebook to compress my CSVs and save them as pickles only. This allows me to reduce my memory usage as I perform the bulk of my work in later notebooks.

import pandas as pd
import pickle

## Compress `rate` CSV

The `rate` CSV file plan-level data on individual rates based on an eligible subscriber’s age, tobacco use, and geographic location, and family-tier rates. Additional information can be found in the [data dictionary provided by CMS here.](https://www.cms.gov/CCIIO/Resources/Data-Resources/Downloads/Rate-Data-Dictionary.pdf)

rate = pd.read_csv('../data/Rate.csv')

rate.info(memory_usage='deep')

rate[rate.select_dtypes(['object']).columns] = rate.select_dtypes(['object']).apply(lambda x: x.astype('category'))

rate.info(memory_usage='deep')

# with open('../pickles/rate.pkl', 'wb') as file:
#     pickle.dump(rate, file)

## Compress `benefits` CSV

The `benefits` CSV file contains plan data on essential health benefits, coverage limits, and cost sharing for each health plan. Additional information can be found in the [data dictionary provided by CMS here.](https://www.cms.gov/CCIIO/Resources/Data-Resources/Downloads/Benefits-and-Cost-Sharing-Data-Dictionary.pdf)

benefits = pd.read_csv('../data/BenefitsCostSharing.csv')

benefits.info(memory_usage='deep')

benefits[benefits.select_dtypes(['object']).columns] = benefits.select_dtypes(['object']).apply(lambda x: x.astype('category'))

benefits.info(memory_usage='deep')

# with open('../pickles/benefits.pkl', 'wb') as file:
#     pickle.dump(benefits, file)
