==> naman_bhandari_capstone/notebooks/03-Vectorize_Plans.ipynb <==
# Vectorize benefits

In this notebook, I imported my cleaned my **rate.pkl** and **benefits_dum.pkl** files and did the following:

- drop certain un-needed columns and filter by 2016
- concatenate my `benefits` dataframe to `rate`  on `'PlanId'` (the plan identifer)
- vectorize the resulting unique list of benefit combinations and save as a matrix

I use the following libraries:

1. `pandas` - A library for manipulating data in a tabular format
2. `pickle` - used to load data previously stored
3. `sqlalchemy` - for interacting with our PostgreSQL database directly from within these notebooks (generating SQL statements).
4. `psycopg2` - a database adapter that sends SQL statements to the actual database.

import pandas as pd
import pickle

import sqlalchemy
from sqlalchemy import create_engine
import psycopg2

import io

Load a few files (and describe them):
- rate.pkl
- benefits_dum.pkl
- crosswalk2.csv

with open('../pickles/rate.pkl', 'rb') as rate:
    rate = pickle.load(rate)
rate.shape

with open('../pickles/benefits_dum.pkl', 'rb') as benefits_dum:
    benefits = pickle.load(benefits_dum)
benefits.shape

crosswalk = pd.read_csv('../data/crosswalk2.csv')
crosswalk.shape

## Cleaning `rate`

Below, I dropped duplicative or un-needed columns.

rate.drop(columns=['IssuerId', 
                   'SourceName',
                   'VersionNum',
                   'ImportDate',
                   'IssuerId2',
                   'FederalTIN'], inplace=True)

I am focusing only on a single plan year, so I'm dropping plans not in the current year.

rate = rate[rate['BusinessYear'] == 2016]

rate_cols = [col for col in rate.columns]

rate.shape

## Cleaning `benefits`

1. Drop `'PlanId'` from `benefits`
2. Rename `'StandardComponentId'` to `'PlanId'`
3. Cast `'PlanId'` as an object (not as a category)

benefits.drop(columns=['PlanId'], inplace=True)
benefits = benefits.rename(columns={'StandardComponentId': 'PlanId'}) 
benefits.PlanId = benefits.PlanId.astype('object')

benefits.shape

## Isolate one-hot encoded dummy columns only in `benefits`

Create a filter to view only the one-hot encoded benefits columns in `benefits`. We will end up with 207 unique benefits.

ben_cols = [x for x in crosswalk['Crosswalk'].unique()]

ben_cols.remove('delete') 

ben_cols.insert(0, 'BenefitName')
ben_cols.insert(1, 'PlanId')

ben_cols_dum = [x for x in ben_cols if x != 'BenefitName' if x != 'PlanId']

benefits[ben_cols_dum].head()

benefits[ben_cols_dum].info()

## Merge `benefits` and `rate` on `'PlanId'`

1. Take `benefits` and groupby `'PlanId'` to get all features of each plan on one row and call the temp dataframe `benefits_planid`.  
    a. We will **sum** our columns in `benefits` when doing the groupby, and then replace any values "greater than 1" with '1'.
2. Then, merge `benefits_planid` to `rate` on `'PlanId'`.

benefits_planid = benefits[ben_cols].drop('BenefitName', axis=1).groupby('PlanId').sum()
benefits_planid[benefits_planid > 1] = 1

ratebenefits = pd.merge(rate, benefits_planid, on='PlanId', how='inner')
ratebenefits[ratebenefits.select_dtypes(['object']).columns] = ratebenefits.select_dtypes(['object']).apply(lambda x: x.astype('category'))

ratebenefits.info()

ratebenefits.shape

Some rows ended up being dropped from `rate` because the `benefits` dataframe did not have all the rows that the `rate` dataframe did. This is because we had filtered on the year 2016 earlier.

## Load and clean `PlanAttributes.csv`

Attributes contains plan-level data on maximum out of pocket payments, deductibles, HSA eligibility, formulary ID, which are useful for users to review.

attributes = pd.read_csv('../data/PlanAttributes.csv')

`attributes` shape before cleaning:

attributes.shape

attributes = attributes[attributes.PlanId.str.contains('-00')]
attributes = attributes[attributes['BusinessYear'] == 2016]

`attributes` shape after cleaning:

attributes.shape

Drop the `'PlanId'` column from `attributes` and rename `'StandardComponentId'` as `'PlanId'`

attributes[attributes.select_dtypes(['object']).columns] = attributes.select_dtypes(['object']).apply(lambda x: x.astype('category'))
attributes.drop(columns=['PlanId'], inplace=True)
attributes = attributes.rename(columns={'StandardComponentId': 'PlanId'}) 
attributes.PlanId = attributes.PlanId.astype('object')

`attributes` shape after dropping:

attributes.shape

Isolate columns that are from `attributes`:

attributes_cols = ['PlanId',
             'IsNoticeRequiredForPregnancy', 
             'IsReferralRequiredForSpecialist', 
             'ChildOnlyOffering', 
             'WellnessProgramOffered', 
             'DiseaseManagementProgramsOffered', 
             'OutOfCountryCoverage', 
             'NationalNetwork']

attributes[attributes_cols].head()

## Merge `ratebenefits` and `attibutes[attibutes_cols]` on `'PlanId'` to create final merged `df` for analysis

df = pd.merge(ratebenefits, attributes[attributes_cols], on='PlanId', how='outer')

df.info()

## Vectorize unique plans and identify with unique `'ben_key'` identifier

df['ben_key'] = df.groupby(ben_cols_dum).ngroup()
ben_key_cols_dum = ben_cols_dum.copy()
ben_key_cols_dum.append('ben_key')

Drop duplicate rows (based on the benefits columns `ben_cols_dum`) to view the vectors that are unique.

no_dupes = df[ben_key_cols_dum].set_index('ben_key').drop_duplicates()

## Upload dataframe to SQL

# set up connection to database (with username/pw if needed)
engine = create_engine('postgres://postgres:postgres@54.190.31.46:5432/')

The below block of code is graciously borrowed from user "mgoldwasser" in this [StackOverflow comment](https://stackoverflow.com/a/45445886/10066207).

def write_to_table(df, db_engine, schema, table_name, if_exists='replace'):
    string_data_io = io.StringIO()
    df.to_csv(string_data_io, sep='|', index=False)
    pd_sql_engine = pd.io.sql.pandasSQL_builder(db_engine, schema=schema)
    table = pd.io.sql.SQLTable(table_name, pd_sql_engine, frame=df,
                               index=False, if_exists=if_exists, schema=schema)
    table.create()
    string_data_io.seek(0)
    string_data_io.readline()  # remove header
    with db_engine.connect() as connection:
        with connection.connection.cursor() as cursor:
            copy_cmd = "COPY %s.%s FROM STDIN HEADER DELIMITER '|' CSV" % (schema, table_name)
            cursor.copy_expert(copy_cmd, string_data_io)
        connection.connection.commit()

# # This function writes my ENTIRE dataframe to a PostgreSQL database hosted on AWS
# write_to_table(df, engine, 'public', 'plans')

In the below cell, I want to remove the one-hot encoded benefits columns and upload a benefits dataframe of approximately 4 million rows and 24 columns, which contains information like State, Age, and Premium. This reduces the size of my table.

Additionally, the benefits columns associated with the plans have been vectorized and identified with foreign keys, making them unnecessary to include in the SQL table.

df_small = df.drop(columns=ben_cols_dum)
df_small = df_small.drop(columns=['RateEffectiveDate', 'RateExpirationDate'])
df_small.shape

# # This function writes df_small, which is the dataframe of unique plans 
# # with select information with one-hot encoded benefit columns removed
# write_to_table(df_small, engine, 'public', 'small')

## Create pickles

# # main CSV of benefits
# df.to_csv('../pickles/df.csv')

# # main dataframe of benefits
# with open('../pickles/df.pkl', 'wb') as file:
#     pickle.dump(df, file)

# # main dataframe of attributes
# with open('../pickles/attributes.pkl', 'wb') as file:
#     pickle.dump(attributes, file)

# # unique vectorized plans
# with open('../pickles/no_dupes.pkl', 'wb') as file:
#     pickle.dump(no_dupes, file)

# # dummy benefits columns
# with open('../pickles/ben_cols_dum.pkl', 'wb') as file:
#     pickle.dump(ben_cols_dum, file)

# # rate columns
# with open('../pickles/rate_cols.pkl', 'wb') as file:
#     pickle.dump(rate_cols, file)
